+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 #### Other Tools
 
 ### Core library changes
+* `dart:async`
+  * Update `Stream.fromIterable` to send a done event after the the error when
+    the iterator's `moveNext` throws, and handle if the `current` getter throws.
+    Issue [33431](http://dartbug.com/33431).
 
 ## 2.1.0-dev.2.0